feat(ui): Add Ctrl+Q shortcut to close the search window
This commit introduces a new keyboard shortcut, Ctrl+Q, to close the search window. Previously, only the Escape key would close the window. This provides an alternative, more conventional shortcut for users familiar with other applications. The keyPressEvent method in the SearchUI class has been modified to handle this new combination of keys.